chore: Fixed references after refactoring
