+++ ra2ce/analyses/configuration/analysis_without_network_config.py
@@ -25,7 +25,7 @@ from __future__ import annotations
 import logging
 from pathlib import Path
 
-from ra2ce.configuration import AnalysisConfigBase, AnalysisIniConfigData
+from ra2ce.analyses.configuration import AnalysisConfigBase, AnalysisIniConfigData
 from ra2ce.graph.network_config_wrapper import NetworkConfigWrapper
