+++ ra2ce/analyses/configuration/readers/analysis_without_network_config_reader.py
@@ -24,14 +24,13 @@ import logging
 from pathlib import Path
 from typing import Optional
 
-from ra2ce.configuration import AnalysisConfigBase
+from ra2ce.analyses.configuration import AnalysisConfigBase
 from ra2ce.analyses.configuration.analysis_ini_config_data import (
     AnalysisWithoutNetworkIniConfigData,
 )
 from ra2ce.analyses.configuration.readers.analysis_config_reader_base import (
     AnalysisConfigReaderBase,
 )
-from ra2ce.graph.network_config_data.network_config_data import NetworkConfigData
 from ra2ce.graph.network_config_data.network_config_data_reader import (
     NetworkConfigDataReader,
 )
