--- ra2ce/configuration/__init__.py
@@ -1,1 +1,0 @@
-from ra2ce.analyses.configuration import AnalysisConfigBase, AnalysisIniConfigData
