+++ ra2ce/runners/indirect_analysis_runner.py
@@ -21,7 +21,7 @@
 
 
 from ra2ce.analyses.indirect import analyses_indirect
-from ra2ce.configuration import AnalysisConfigBase
+from ra2ce.analyses.configuration import AnalysisConfigBase
 from ra2ce.configuration.config_wrapper import ConfigWrapper
 from ra2ce.runners.analysis_runner_protocol import AnalysisRunner
 
@@ -32,7 +32,10 @@ class IndirectAnalysisRunner(AnalysisRunner):
 
     @staticmethod
     def can_run(ra2ce_input: ConfigWrapper) -> bool:
-        return not ra2ce_input.analysis_config or "indirect" in ra2ce_input.analysis_config.config_data
+        return (
+            not ra2ce_input.analysis_config
+            or "indirect" in ra2ce_input.analysis_config.config_data
+        )
 
     def run(self, analysis_config: AnalysisConfigBase) -> None:
         analyses_indirect.IndirectAnalyses(
