+++ tests/configuration/test_config_wrapper.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 import pytest
 
-from ra2ce.configuration import AnalysisConfigBase
+from ra2ce.analyses.configuration import AnalysisConfigBase
 from ra2ce.configuration.config_wrapper import ConfigWrapper
 from ra2ce.graph.network_config_wrapper import NetworkConfigWrapper
 from tests import test_data
